requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-filter==2.0.0
django-fontawesome==1.0
-e git+https://github.com/JoabMendes/django-geoposition@ab7dac066bf61a5b2ec18a30c7977a4ffc6d568b#egg=django_geoposition
django-rest-recaptcha==1.0.2
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.9.0
idna==2.7
Markdown==2.6.11
mccabe==0.6.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
urllib3==1.23
whitenoise==4.0
